Add files via upload
Exploring San Francisco city employee salary data with pandas, matplotlib, and seaborn. 

This data reveals that job titles related to transit play a major role in San Francisco's economy. Transit Operator turned out to be the most common job title in the data between 2011 and 2014. This, coupled with the highest salary belonging to the General Manager of the transit authority, paint an optimistic picture of San Francisco's priorities. Such heavy investment in transit emphasizes the government's commitment to addressing concerns such as pollution, traffic congestion, and high population density.